requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
kaggle==1.5.16
python-dateutil==2.8.2
python-slugify==8.0.1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tqdm==4.66.1
urllib3==2.0.4
webencodings==0.5.1